refactor(tailer): split TailAndProcess/processParsedEvent into cohesive helpers
CodeScene flagged core/pkg/tailer/tailer.go (Code Health 5.34/10, friction
rising 0.24→0.34 over 75 revisions, #806). Root cause: TailAndProcess and
processParsedEvent had grown into two "God functions" — every adapter
feature added over 75 revisions (task tracking, background-process
tracking, interrupt/denial flags, task estimate/summary/question markers)
bolted onto the same two functions instead of getting its own home, so
comprehension cost compounded on the shared hot path every adapter's
transcript parsing goes through.

No adapter-specific parsing had leaked into the shared tailer — format
parsing is already properly delegated via the TranscriptParser interface,
and adapter-specific model-config lookups already live in
tailer_config.go — so this is a pure structural cleanup, not a layering
fix.

Changes (behavior-preserving, no logic changes — only extraction):
- TailAndProcess decomposed into scanNewLines (the line-scan loop),
  parseTranscriptLine, applySkippedEvent, and resetAccumulatorsForRotation,
  leaving a ~45-line orchestrator.
- processParsedEvent decomposed into applyToolCallDeltas, applyTaskDeltas,
  applyBackgroundProcessDeltas, sweepOpenToolCallsOnTurnDone, and
  updateInterruptAndDenialFlags/applyAssistantTextAndMarkers, each a single
  cohesive concern, called in the original order.
- readLineCapped (+ its sentinel errors/const) moved to line_reader.go;
  GetLedgerState/SetLedgerState moved to tailer_ledger.go — mirrors the
  existing tailer_metrics.go/tailer_config.go split-by-concern convention.

Verified behavior-preserving: go test ./core/... -race -count=1 green,
tools/replay-fixtures.sh green (141 fixtures byte-identical across every
adapter), tools/preflight.sh --only go green.

Closes #806
